requirements.txt
astroid==2.11.5
certifi==2021.10.8
cffi==1.15.1
chardet==5.1.0
Click
cryptography==39.0.0
docusign-esign==3.22.0
docusign-rooms==1.1.0
docusign-monitor==1.1.0
docusign-click==1.2.2
docusign-admin==1.3.0
Flask==1.1.1
Flask-OAuthlib==0.9.6
Flask-Session
flask-wtf==0.14.3
flake8==5.0.4
idna==3.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2>=2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.7.0
oauthlib==2.1.0
pipenv==2022.1.8
py-oauth2==0.0.10
pycparser==2.21
pylint==2.14.1
python-dateutil==2.8.2
python-dotenv==0.21.0
requests>=2.28.2
requests-oauthlib==1.1.0
six==1.13.0
urllib3>=1.26.14
virtualenv==20.21.0
virtualenv-clone==0.5.7
Werkzeug==2.0.0
wrapt==1.11.2
